More level updates
Added UI arrows to guide the player. Kevin-proofed the level by adding invisible walls around the outside of the track. Slightly increased the movement speed of the AI. Add a trigger/timer system for fireworks when the player crosses the finish line. Added a small scene to Scott F's title screen.